delete user dialog functionality

+++ src/store/apis/users.ts
@@ -1,5 +1,6 @@
 import axios, { AxiosResponse } from 'axios';
 import { secretIp } from '../../../secrets/secretStuff';
+import { extractId } from '../actions/helper_functions';
 
 export function getAllUsers(): Promise<AxiosResponse> {
   const accessToken = localStorage.getItem('access_token');
@@ -30,9 +31,11 @@ export function getOneUser(userId: string): Promise<AxiosResponse> {
 
 export function deleteUser(userId: string): Promise<AxiosResponse> {
   const accessToken = localStorage.getItem('access_token');
-  return axios.delete(`${secretIp}/api/users/${userId}`, {
-    headers: {
-      authorization: accessToken,
-    },
-  });
+  const loggedInUser = extractId(accessToken);
+  if (loggedInUser.toString() !== userId)
+    return axios.delete(`${secretIp}/api/users/${userId}`, {
+      headers: {
+        authorization: accessToken,
+      },
+    });
 }
